[#176236216] - Split weights off into submodule
